requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-environ==0.11.2
django-import-export==3.3.8
django-js-asset==2.2.0
et-xmlfile==1.1.0
kombu==5.4.2
MarkupPy==1.14
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.5
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
